requirements.txt
tweepy
textblob
mysql-connector-python
python-dateutil